Added user to projects releationship

+++ app/Http/Controllers/ProjectController.php
@@ -4,11 +4,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Project;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
     public function index() {
-        $projects = Project::all();
+        $projects = Project::where('user_id', Auth::id())
+            ->get();
 
         return view('account.projects.index', compact('projects'));
     }
@@ -19,7 +21,10 @@ class ProjectController extends Controller
     
     public function store(Request $request) {
         $project = new Project();
-        $project::create($request->all());
+        $project::create([
+            "title"   => $request->title,
+            "user_id" => Auth::id()
+        ]);
 
         return redirect('account/projects');
     }
@@ -37,16 +42,22 @@ class ProjectController extends Controller
     }
     
     public function update(Request $request, $id) {
-        Project::where('id', $id)->update([
-            "title" => $request->title
-        ]);
+        Project::where('id', $id)
+            ->where('user_id', Auth::id())
+            ->update([
+                "title" => $request->title
+            ]);
 
         return back();
     }
     
     public function destroy($id) {
         $project = Project::where('id', $id)->first();
-        $project->deleteRelated();
+
+        if ($project->user_id == Auth::id()) {
+            $project->deleteRelated();
+            return redirect('account/projects');
+        } else return redirect('account/projects');
 
         return redirect('account/projects');
     }
